requirements.txt
fastapi
uvicorn
psycopg2-binary
boto3
httpx
requests
pymodbus
cryptography
ipfshttpclient
